make changes to type classes, methods.
