+++ src/main/java/core/mate/academy/model/BulldozerProducer.java
@@ -1,11 +1,15 @@
 package core.mate.academy.model;
 
 import core.mate.academy.service.MachineProducer;
+import java.util.ArrayList;
 import java.util.List;
 
 public class BulldozerProducer implements MachineProducer<Bulldozer> {
     @Override
-    public List<? extends Machine> get() {
-        return new Bulldozer().get();
+    public List<Bulldozer> get() {
+        List<Bulldozer> bulldozers = new ArrayList<>();
+        bulldozers.add(new Bulldozer("first", 1));
+        bulldozers.add(new Bulldozer("second", 2));
+        return bulldozers;
     }
 }
