+++ src/main/java/core/mate/academy/model/Excavator.java
@@ -1,14 +1,10 @@
 package core.mate.academy.model;
 
-import core.mate.academy.service.MachineProducer;
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * Add some custom fields that could be only in Excavator
  * Do not remove no-args constructor
  */
-public class Excavator extends Machine implements MachineProducer<Excavator> {
+public class Excavator extends Machine {
     public Excavator() {
     }
 
@@ -19,13 +15,5 @@ public class Excavator extends Machine implements MachineProducer<Excavator> {
     @Override
     public void doWork() {
         System.out.println("Excavator started to work");
-    }
-
-    @Override
-    public List<Excavator> get() {
-        List<Excavator> excavators = new ArrayList<>();
-        excavators.add(new Excavator("Third", 3));
-        excavators.add(new Excavator("Four", 4));
-        return excavators;
     }
 }
