+++ src/main/java/core/mate/academy/model/ExcavatorProducer.java
@@ -1,11 +1,15 @@
 package core.mate.academy.model;
 
 import core.mate.academy.service.MachineProducer;
+import java.util.ArrayList;
 import java.util.List;
 
 public class ExcavatorProducer implements MachineProducer<Excavator> {
     @Override
-    public List<? extends Machine> get() {
-        return new Excavator().get();
+    public List<Excavator> get() {
+        List<Excavator> excavators = new ArrayList<>();
+        excavators.add(new Excavator("Third", 3));
+        excavators.add(new Excavator("Four", 4));
+        return excavators;
     }
 }
