+++ src/main/java/core/mate/academy/model/Truck.java
@@ -1,14 +1,10 @@
 package core.mate.academy.model;
 
-import core.mate.academy.service.MachineProducer;
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * Add some custom fields that could be only in Truck
  * Do not remove no-field constructor
  */
-public class Truck extends Machine implements MachineProducer<Truck> {
+public class Truck extends Machine {
     public Truck() {
     }
 
@@ -19,13 +15,5 @@ public class Truck extends Machine implements MachineProducer<Truck> {
     @Override
     public void doWork() {
         System.out.println("Truck started to work");
-    }
-
-    @Override
-    public List<Truck> get() {
-        List<Truck> trucks = new ArrayList<>();
-        trucks.add(new Truck("Five", 5));
-        trucks.add(new Truck("Six", 6));
-        return trucks;
     }
 }
