+++ src/main/java/core/mate/academy/model/TruckProducer.java
@@ -1,11 +1,15 @@
 package core.mate.academy.model;
 
 import core.mate.academy.service.MachineProducer;
+import java.util.ArrayList;
 import java.util.List;
 
 public class TruckProducer implements MachineProducer<Truck> {
     @Override
-    public List<? extends Machine> get() {
-        return new Truck().get();
+    public List<Truck> get() {
+        List<Truck> trucks = new ArrayList<>();
+        trucks.add(new Truck("Five", 5));
+        trucks.add(new Truck("Six", 6));
+        return trucks;
     }
 }
