+++ src/main/java/core/mate/academy/service/MachineProducer.java
@@ -3,10 +3,10 @@ package core.mate.academy.service;
 import core.mate.academy.model.Machine;
 import java.util.List;
 
-public interface MachineProducer<T> {
+public interface MachineProducer<T extends Machine> {
     /**
      * In your implementations - create 2-3 models, add them to the list and return
      * @return - the list of models
      */
-    List<? extends Machine> get();
+    List<T> get();
 }
