+++ src/main/java/core/mate/academy/service/MachineService.java
@@ -6,7 +6,7 @@ import java.util.List;
 /**
  * Parameterize this service and add its implementation in a separate class.
  */
-public interface MachineService<T> {
+public interface MachineService<T extends Machine> {
 
     /**
      * Return the list of machines.
@@ -17,7 +17,7 @@ public interface MachineService<T> {
      * @param type - any class of Machine sub class. For example: Truck.class or Bulldozer.class
      * @return the list of machines
      */
-    List<Machine> getAll(Class<? extends Machine> type);
+    List<Machine> getAll(Class<? extends T> type);
 
     /**
      * Fill the machines list with passed value
@@ -27,7 +27,7 @@ public interface MachineService<T> {
      * @param machines - list of machines to be filled with value
      * @param value    - any object of machine sub class
      */
-    void fill(List<? super Machine> machines, Machine value);
+    void fill(List<? super T> machines, T value);
 
     /**
      * Call the method doWork() from each machine.
@@ -35,5 +35,5 @@ public interface MachineService<T> {
      *
      * @param machines - the list of machines
      */
-    void startWorking(List<? extends Machine> machines);
+    void startWorking(List<? extends T> machines);
 }
