+++ src/main/java/core/mate/academy/service/MachineServiceImpl.java
@@ -15,7 +15,7 @@ import java.util.List;
  */
 public class MachineServiceImpl implements MachineService<Machine> {
     @Override
-    public List<Machine> getAll(Class type) {
+    public List<Machine> getAll(Class<? extends Machine> type) {
         if (type == Bulldozer.class) {
             BulldozerProducer bulldozerProducer = new BulldozerProducer();
             List<? extends Machine> bulldozer = bulldozerProducer.get();
